+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - irc: add option `addreplace` in commands `/ignore`, `/notify` and `/server` ([#2095](https://github.com/weechat/weechat/issues/2095))
 - relay: add option `addreplace` in commands `/relay` and `/remote` ([#2095](https://github.com/weechat/weechat/issues/2095))
 - core: allow buffer id in commands `/buffer` and `/print`, in addition to number and name
+- core: allow mask in command `/item refresh`
 - relay: enable websocket extension "permessage-deflate" with "api" relay only ([#1549](https://github.com/weechat/weechat/issues/1549))
 - relay/weechat: add line id in buffer lines sent to clients
 - api: allow NULL value for key in hashtable